fix part of code smells
